Connected database to colors2...
Connected mysql faure maraidb table colors to colors2 page. The drop down options are now the name column of the db table colors. If a new user visits the colors2 page, the drop downs are the default colors. If a current user changes there colors, the changes are reflected in the drop down.

Im not sure how to connect the hex_values yet. But added some php to pull the hex values and put them into a hex_values array. Hopefully it will be easy to connect.

Changed colors2 Number of Colors error check logic. If the user has less then 10 colors in there data base, then they the max_colors is the number of name values for their user_id. If the number of colors is greater then 10, then max_colors equals 10.

Added session to colors2 so that the users user_id can be gotten.